src/requirementsv4.txt
absl-py==2.1.0
alembic==1.13.3
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
azure-core==1.31.0
azure-storage-blob==12.23.1
beautifulsoup4==4.12.3
black==24.8.0
blinker==1.8.2
boto3==1.35.29
botocore==1.35.29
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
Cython==3.0.11
databricks-sdk==0.33.0
decorator==5.1.1
Deprecated==1.2.14
# detectron2==0.6
docker==7.1.0
environs==9.5.0
et-xmlfile==1.1.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.54.1
fsspec==2024.9.0
funcy==2.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.56.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
img2table==1.3.0
imgaug==0.4.0
importlib_metadata==8.4.0
iopath==0.1.9
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
layoutparser==0.3.4
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
minio==7.2.9
mlflow==2.16.2
mlflow-skinny==2.16.2
mmh3==3.0.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
opencv-contrib-python-headless==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt-einsum==3.3.0
packaging==24.1
paddlepaddle==2.6.2
pandas==2.2.3
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
platformdirs==4.3.6
polars==1.8.2
portalocker==2.10.1
protobuf==3.20.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post5
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.20.0
pymilvus==2.4.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
# pywin32==306
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.1
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.20.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.1.0
urllib3==2.2.3
waitress==3.0.0
Werkzeug==3.0.4
wrapt==1.16.0
XlsxWriter==3.2.0
yacs==0.1.8
zipp==3.20.2
